[LangCodeExpander] Speed up region codes lookups
