+++ xbmc/utils/i18n/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(HEADERS TableISO3166_1.h)
+
+if(NOT ENABLE_STATIC_LIBS)
+  core_add_library(utils_i18n)
+endif()
